+++ src/main/java/com/pbs/ams/web/mappers/AmsTradeOrderMapper.java
@@ -3,13 +3,13 @@ package com.pbs.ams.web.mappers;
 import com.pbs.ams.web.model.AmsTradeOrder;
 
 public interface AmsTradeOrderMapper {
-    int deleteByPrimaryKey(Integer tradeOrderId);
+    int deleteByPrimaryKey(Long tradeOrderId);
 
     int insert(AmsTradeOrder record);
 
     int insertSelective(AmsTradeOrder record);
 
-    AmsTradeOrder selectByPrimaryKey(Integer tradeOrderId);
+    AmsTradeOrder selectByPrimaryKey(Long tradeOrderId);
 
     int updateByPrimaryKeySelective(AmsTradeOrder record);
